pyproject.toml
[project]
name = "kratos-keto-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]
